move functions tests into `test/unit/` because that is what they are
