+++ test/unit/api-health.test.js
@@ -1,7 +1,7 @@
 import test from "ava";
 
-import { handler, healthLog } from "../netlify/functions/health/index.js";
-import { mockLog } from "./mocks.js";
+import { handler, healthLog } from "../../netlify/functions/health/index.js";
+import { mockLog } from "../mocks.js";
 
 test("GET /api/health handler", async (t) => {
   const logs = mockLog(healthLog);
